+++ src/submissions/tools/scripts/hello.py
@@ -1,5 +1,8 @@
 """
 Test script for startup_scripts
 """
-def script(ctx):
+from .. import register_script
+
+@register_script
+def hello(ctx):
     print("\n\nHello! Welcome to Robotics Submission Tracker.\n\n")